# Faster require