--- a/spec/fake_stripe/utils_spec.rb
+++ b/spec/fake_stripe/utils_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+describe FakeStripe::Utils do
+  describe '#find_available_port' do
+    it 'returns a port value' do
+      expect(described_class.find_available_port).to be_a Numeric
+    end
+  end
+end