@samcday 2017-09-04
@odlp it looks like this (awesome) PR never ended up in a release anywhere. Would you mind cutting a new release to RubyGems sometime?