@odlp 2017-09-05
@samcday good spot - I just released `v0.1.0` to Rubygems: https://rubygems.org/gems/fake_stripe/versions/0.1.0